Refactoring and fix checkstyle error
