+++ src/test/java/cs/f10/t1/nursetraverse/autocomplete/AssociableWordTest.java
@@ -4,6 +4,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.LinkedList;
 
+import cs.f10.t1.nursetraverse.model.autocomplete.AssociableWord;
+import cs.f10.t1.nursetraverse.model.autocomplete.CommandWord;
 import org.junit.jupiter.api.Test;
 
 class AssociableWordTest {
