+++ src/test/java/cs/f10/t1/nursetraverse/autocomplete/AutoCompleteWordStorageTest.java
@@ -6,6 +6,8 @@ import static cs.f10.t1.nursetraverse.model.util.SampleDataUtil.getVisitTodos;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import cs.f10.t1.nursetraverse.model.appointment.AutoCompleteWord;
+import cs.f10.t1.nursetraverse.model.autocomplete.ObjectWord;
 import org.junit.jupiter.api.Test;
 
 import cs.f10.t1.nursetraverse.model.HistoryRecord;
