+++ src/test/java/cs/f10/t1/nursetraverse/autocomplete/CommandWordTest.java
@@ -3,6 +3,7 @@ package cs.f10.t1.nursetraverse.autocomplete;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import cs.f10.t1.nursetraverse.model.autocomplete.CommandWord;
 import org.junit.jupiter.api.Test;
 
 class CommandWordTest {
