+++ src/test/java/cs/f10/t1/nursetraverse/autocomplete/PrefixWordTest.java
@@ -2,6 +2,7 @@ package cs.f10.t1.nursetraverse.autocomplete;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import cs.f10.t1.nursetraverse.model.autocomplete.PrefixWord;
 import org.junit.jupiter.api.Test;
 
 class PrefixWordTest {
